+++ Cargo.toml
@@ -60,6 +60,7 @@ ic-management-canister-types = { version = "0.7.1" }
 ic-utils = { version = "0.47.0" }
 icp = { path = "crates/icp" }
 icp-canister-interfaces = { path = "crates/icp-canister-interfaces" }
+icp-deploy-canister = { path = "crates/icp-deploy-canister" }
 icp-sync-plugin = { path = "crates/icp-sync-plugin" }
 ic-identity-hsm = "0.47.0"
 icrc-ledger-types = "0.1.10"
