+++ crates/icp-cli/src/operations/mod.rs
@@ -1,3 +1,4 @@
+pub(crate) mod access;
 pub(crate) mod binding_env_vars;
 pub(crate) mod build;
 pub(crate) mod bundle;
