+++ crates/icp-cli/src/operations/sync.rs
@@ -1,15 +1,22 @@
+use async_trait::async_trait;
 use candid::Principal;
 use futures::{StreamExt, stream::FuturesOrdered};
 use ic_agent::Agent;
 use icp::{
     Canister,
-    canister::sync::{Params, Synchronize, SynchronizeError},
+    canister::sync::{Synchronize, SynchronizeError},
     package::PackageCache,
     prelude::PathBuf,
 };
+use icp_deploy_canister::manifest::adapter::prebuilt::SourceField;
+use icp_deploy_canister::sync_exec::{
+    PluginExecutor, PluginExecutorError, PluginInvocation, ScriptInvocation, StepProgress,
+};
+use icp_deploy_canister::{SyncCanisterError, SyncStepContext, run_sync_steps};
 use snafu::prelude::*;
 use std::collections::BTreeMap;
 use std::sync::Arc;
+use tokio::sync::Mutex;
 use tracing::error;
 
 use crate::progress::{MultiStepProgressBar, ProgressManager, ProgressManagerSettings};
@@ -24,14 +31,102 @@ pub struct SyncOperationError {
 struct SyncFailure {
     canister_name: String,
     canister_id: Principal,
-    error: SynchronizeError,
+    error: SyncCanisterError,
     progress_output: Vec<String>,
 }
 
-/// Synchronizes a single canister using its configured sync steps
+/// Per-canister mutable state guarded so the `&self` [`PluginExecutor`] can drive
+/// the (mutable, sequential) progress bar.
+struct SyncStepState<'a> {
+    pb: &'a mut MultiStepProgressBar,
+    /// 1-based index of the step about to run, for the progress header.
+    next: usize,
+}
+
+/// [`PluginExecutor`] that runs a resolved step via the host [`Synchronize`]
+/// implementation (WASI plugin / subprocess script) and frames it on the
+/// canister's multi-step progress bar. The library owns the step loop and all
+/// input derivation ([`run_sync_steps`]); this only performs the host action and
+/// streams its output.
+struct AgentSyncExecutor<'a> {
+    syncer: Arc<dyn Synchronize>,
+    agent: Agent,
+    pkg_cache: &'a PackageCache,
+    total: usize,
+    state: Mutex<SyncStepState<'a>>,
+}
+
+impl AgentSyncExecutor<'_> {
+    /// Frame a step on the shared progress bar: advance the counter, print the
+    /// header, run `f` against a fresh line sender, and close the step. Holding
+    /// the guard across `f` keeps steps framed sequentially on the shared bar.
+    async fn framed<F, Fut>(
+        &self,
+        header: impl FnOnce(usize, usize) -> String,
+        f: F,
+    ) -> Result<Vec<String>, SynchronizeError>
+    where
+        F: FnOnce(tokio::sync::mpsc::Sender<String>) -> Fut,
+        Fut: Future<Output = Result<Vec<String>, SynchronizeError>>,
+    {
+        let mut st = self.state.lock().await;
+        st.next += 1;
+        let header = header(st.next, self.total);
+        let tx = st.pb.begin_step(header);
+        let result = f(tx).await;
+        st.pb.end_step().await;
+        result
+    }
+}
+
+#[async_trait]
+impl PluginExecutor for AgentSyncExecutor<'_> {
+    async fn run_plugin(
+        &self,
+        invocation: PluginInvocation,
+        _progress: Option<&dyn StepProgress>,
+    ) -> Result<Vec<String>, PluginExecutorError> {
+        let src = match &invocation.source {
+            SourceField::Local(l) => format!("path: {}", l.path),
+            SourceField::Remote(r) => format!("url: {}", r.url),
+        };
+        self.framed(
+            |n, total| format!("\nSyncing: plugin {src} {n} of {total}"),
+            |tx| async move {
+                self.syncer
+                    .run_plugin(&invocation, &self.agent, Some(tx), self.pkg_cache)
+                    .await
+            },
+        )
+        .await
+        .map_err(|source| PluginExecutorError::Plugin {
+            source: Box::new(source),
+        })
+    }
+
+    async fn run_script(
+        &self,
+        invocation: ScriptInvocation,
+        _progress: Option<&dyn StepProgress>,
+    ) -> Result<Vec<String>, PluginExecutorError> {
+        let desc = invocation.commands.join("\n");
+        self.framed(
+            |n, total| format!("\nSyncing: script {desc} {n} of {total}"),
+            |tx| async move { self.syncer.run_script(&invocation, Some(tx)).await },
+        )
+        .await
+        .map_err(|source| PluginExecutorError::Script {
+            source: Box::new(source),
+        })
+    }
+}
+
+/// Synchronize a single canister's steps through the library, framing progress
+/// on `pb`. Environment variables are applied separately by the caller.
+#[allow(clippy::too_many_arguments)]
 async fn sync_canister(
-    syncer: &Arc<dyn Synchronize>,
-    agent: &Agent,
+    syncer: Arc<dyn Synchronize>,
+    agent: Agent,
     canister_path: PathBuf,
     canister_id: Principal,
     canister_info: &Canister,
@@ -41,45 +136,27 @@ async fn sync_canister(
     proxy: Option<Principal>,
     pb: &mut MultiStepProgressBar,
     pkg_cache: &PackageCache,
-) -> Result<Vec<String>, SynchronizeError> {
-    let step_count = canister_info.sync.steps.len();
-    let mut stderr_lines = Vec::new();
-
-    for (i, step) in canister_info.sync.steps.iter().enumerate() {
-        // Indicate to user the current step being executed
-        let current_step = i + 1;
-        let pb_hdr = format!("\nSyncing: {step} {current_step} of {step_count}");
-
-        let tx = pb.begin_step(pb_hdr);
-
-        // Execute step
-        let sync_result = syncer
-            .sync(
-                step,
-                &Params {
-                    path: canister_path.clone(),
-                    cid: canister_id,
-                    environment: environment.to_owned(),
-                    network: network.to_owned(),
-                    canister_ids: canister_ids.clone(),
-                    proxy,
-                },
-                agent,
-                Some(tx),
-                pkg_cache,
-            )
-            .await;
-
-        // Ensure background receiver drains all messages
-        pb.end_step().await;
-
-        stderr_lines.extend(sync_result?);
-    }
-
-    Ok(stderr_lines)
+) -> Result<Vec<String>, SyncCanisterError> {
+    let ctx = SyncStepContext {
+        canister_path,
+        canister_id,
+        environment: environment.to_owned(),
+        network: network.to_owned(),
+        canister_ids: canister_ids.clone(),
+        proxy,
+    };
+    let executor = AgentSyncExecutor {
+        syncer,
+        agent,
+        pkg_cache,
+        total: canister_info.sync.steps.len(),
+        state: Mutex::new(SyncStepState { pb, next: 0 }),
+    };
+    run_sync_steps(canister_info, &ctx, &executor, None).await
 }
 
 /// Orchestrates syncing multiple canisters with progress tracking
+#[allow(clippy::too_many_arguments)]
 pub(crate) async fn sync_many(
     syncer: Arc<dyn Synchronize>,
     agent: Agent,
@@ -105,10 +182,9 @@ pub(crate) async fn sync_many(
             let canister_ids = canister_ids.clone();
 
             async move {
-                // Define the sync logic
                 let sync_result = sync_canister(
-                    &syncer,
-                    &agent,
+                    syncer,
+                    agent,
                     canister_path,
                     cid,
                     &canister_info,
