+++ crates/icp-deploy-canister/Cargo.toml
@@ -1,0 +1,50 @@
+[package]
+name = "icp-deploy-canister"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+publish.workspace = true
+
+# This crate is intentionally dependency-light so that its install/sync core and
+# project model can be compiled into a canister. All host-only IO (filesystem,
+# HTTP, the ICP API, sync-step execution, the canister-id store) is abstracted
+# behind trait objects; do NOT add ic-agent, reqwest, tokio, wasmtime, keyring,
+# bollard, sysinfo, or std::fs-driven crates here.
+
+[features]
+# Enables `clap::ValueEnum` derives on manifest enums used as CLI value types
+# (e.g. `ArgsFormat`). Enabled transitively by `icp/clap`.
+clap = ["dep:clap"]
+
+[dependencies]
+async-trait.workspace = true
+bigdecimal.workspace = true
+camino.workspace = true
+candid.workspace = true
+clap = { workspace = true, optional = true }
+candid_parser.workspace = true
+futures.workspace = true
+glob.workspace = true
+hex.workspace = true
+ic-management-canister-types.workspace = true
+indexmap.workspace = true
+itertools.workspace = true
+num-bigint.workspace = true
+num-integer.workspace = true
+num-traits.workspace = true
+pathdiff.workspace = true
+schemars.workspace = true
+serde.workspace = true
+sha2.workspace = true
+serde_json.workspace = true
+serde_yaml.workspace = true
+snafu.workspace = true
+strum.workspace = true
+tracing.workspace = true
+url.workspace = true
+
+[dev-dependencies]
+camino-tempfile.workspace = true
+indoc.workspace = true
+jsonschema.workspace = true
+tokio.workspace = true
