+++ crates/icp-deploy-canister/src/manifest/adapter/mod.rs
