+++ crates/icp-deploy-canister/src/manifest/adapter/plugin.rs
