+++ crates/icp-deploy-canister/src/manifest/adapter/prebuilt.rs
