+++ crates/icp-deploy-canister/src/manifest/adapter/script.rs
