+++ crates/icp-deploy-canister/src/manifest/canister.rs
