+++ crates/icp-deploy-canister/src/manifest/dependency.rs
