+++ crates/icp-deploy-canister/src/manifest/environment.rs
