+++ crates/icp-deploy-canister/src/manifest/network.rs
