+++ crates/icp-deploy-canister/src/manifest/project.rs
