+++ crates/icp-deploy-canister/src/manifest/recipe.rs
