+++ crates/icp-deploy-canister/src/manifest/serde_helpers.rs
