+++ crates/icp/src/canister/sync/plugin.rs
@@ -1,13 +1,11 @@
 use camino::Utf8PathBuf;
-use candid::Principal;
 use ic_agent::Agent;
+use icp_deploy_canister::sync_exec::PluginInvocation;
 use icp_sync_plugin::{RunPluginError, run_plugin};
 use snafu::prelude::*;
 use tokio::sync::mpsc::Sender;
 
-use crate::{canister::wasm, manifest::adapter::plugin::Adapter, package::PackageCache};
-
-use super::Params;
+use crate::{canister::wasm, package::PackageCache};
 
 #[derive(Debug, Snafu)]
 pub enum PluginError {
@@ -21,12 +19,13 @@ pub enum PluginError {
     Run { source: RunPluginError },
 }
 
-pub(super) async fn sync(
-    adapter: &Adapter,
-    params: &Params,
+/// Fetch and run a WASI plugin against a canister for a fully-resolved
+/// [`PluginInvocation`]. Dispatch and input derivation happen in
+/// `icp-deploy-canister`; this only performs the host-only wasm resolution and
+/// wasmtime execution.
+pub(super) async fn run(
+    invocation: &PluginInvocation,
     agent: &Agent,
-    environment: &str,
-    proxy: Option<Principal>,
     stdio: Option<Sender<String>>,
     pkg_cache: &PackageCache,
 ) -> Result<Vec<String>, PluginError> {
@@ -35,21 +34,18 @@ pub(super) async fn sync(
     //    - Remote: downloaded to cache (sha256 required, enforced at parse time) and the
     //      stable cache path is returned — no temp file needed.
     let wasm_path = wasm::resolve(
-        &adapter.source,
-        &params.path,
-        adapter.sha256.as_deref(),
+        &invocation.source,
+        &invocation.base_dir,
+        invocation.sha256.as_deref(),
         stdio.as_ref(),
         pkg_cache,
     )
     .await?;
 
-    // 2. Collect inputs as manifest strings. `run_plugin` preopens the `dirs`
-    //    and reads the `files` itself — both anchored at `base_dir`, and both
-    //    subject to the runtime's path-safety checks (no escaping or symlinked
-    //    paths).
-    let base_dir = Utf8PathBuf::from(params.path.as_str());
-    let dirs: Vec<String> = adapter.dirs.clone().unwrap_or_default();
-    let files: Vec<String> = adapter.files.clone().unwrap_or_default();
+    // 2. `run_plugin` preopens the `dirs` and reads the `files` itself — both
+    //    anchored at `base_dir`, and both subject to the runtime's path-safety
+    //    checks (no escaping or symlinked paths).
+    let base_dir = Utf8PathBuf::from(invocation.base_dir.as_str());
 
     // 3. Run the plugin (blocking call — signal Tokio that this thread will block).
     let identity_principal = agent
@@ -57,7 +53,11 @@ pub(super) async fn sync(
         .map_err(|err| PluginError::GetIdentityPrincipal { err })?;
 
     let agent_clone = agent.clone();
-    let environment_owned = environment.to_owned();
+    let dirs = invocation.dirs.clone();
+    let files = invocation.files.clone();
+    let cid = invocation.canister_id;
+    let proxy = invocation.proxy;
+    let environment_owned = invocation.environment.clone();
     let stdio_clone = stdio.clone();
 
     tokio::task::block_in_place(|| {
@@ -66,7 +66,7 @@ pub(super) async fn sync(
             base_dir,
             dirs,
             files,
-            params.cid,
+            cid,
             agent_clone,
             proxy,
             identity_principal,
